added unique indexes for login and email and implemented MongoServerError for duplicate keys

+++ src/domain/auth-service.ts
@@ -10,7 +10,8 @@ import { createConfirmationCode } from "../utils/auth-utils/create-user-confirma
 import { createCodeExpirationDate } from "../utils/auth-utils/create-code-expiration-date";
 
 export const authService = {
-  async registerNewUser(body: UserInputModel): Promise<boolean> {
+  async registerNewUser(body: UserInputModel): Promise<TFieldError[]> {
+    const errors: TFieldError[] = [];
     const { login, email, password } = body;
     const passwordSalt = await bcrypt.genSalt(10);
     const passwordHash = await usersService._generateHash(
@@ -32,14 +33,26 @@ export const authService = {
       },
     };
     const createUser = await usersCommandsRepository.createNewUser(newUser);
-    try {
-      await emailManager.sendEmail(newUser);
-      return true;
-    } catch (error) {
-      console.error(error);
-      await usersCommandsRepository.deleteUser(createUser.id);
-      return false;
+    if (!createUser) {
+      errors.push({
+        message:
+          "User with the given email or login already exists",
+        field: "registration",
+      });
+    } else {
+      try {
+        await emailManager.sendEmail(newUser);
+      } catch (error) {
+        console.error(error);
+        await usersCommandsRepository.deleteUser(createUser.id);
+        errors.push({
+          message:
+            "Something went wrong with registration/ User was not created",
+          field: "registration",
+        });
+      }
     }
+    return errors;
   },
   async confirmCode(code: string): Promise<TFieldError[]> {
     const errors: TFieldError[] = [];
